backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
certifi
cffi
charset-normalizer
cryptography
distro
Django
django-cors-headers
djangorestframework
et_xmlfile
frozenlist
google-auth
google-genai
h11
httpcore
httpx
idna
joblib
multidict
numpy
openpyxl
pandas
propcache
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
python-dateutil
python-dotenv
requests
rsa
scikit-learn
scipy
six
sniffio
sqlparse
tenacity
threadpoolctl
typing-inspection
typing_extensions
tzdata
ucimlrepo
urllib3
websockets
yarl
